pyApp/app/requirements.txt
fastapi
uvicorn
mysql-connector